+++ fake-backend/serve.js
@@ -47,12 +47,15 @@ app.get('/api/v1/data/contractagreements', (req, res) => {
   res.json(contractAgreements);
 });
 
-app.get('/catalogs/1', (req, res) => {
-  res.json(catalog1);
-});
-
-app.get('/catalogs/2', (req, res) => {
-  res.json(catalog2);
+app.get('/api/v1/data/catalog', (req, res) => {
+  let providerUrl = req.query.providerUrl;
+  if (providerUrl === 'http://existing-other-connector/v1/ids/data') {
+    res.json(catalog1);
+  } else if (providerUrl === 'https://test.gg') {
+    res.json(catalog2);
+  } else {
+    res.json({contractOffers: []});
+  }
 });
 
 app.listen(3000, function () {