+++ src/modules/edc-demo/services/catalog-api-url.service.ts
@@ -1,6 +1,7 @@
 import {Injectable} from '@angular/core';
 import {AppConfigService} from '../../app/config/app-config.service';
 
+
 /**
  * Stores Catalog API URLs
  */
@@ -11,15 +12,15 @@ export class CatalogApiUrlService {
   /**
    * From app-config.json, not user editable
    */
-  private readonly presetCatalogApiUrls = new Array<string>();
+  private readonly presetProviders = new Array<string>();
 
   /**
    * User-editable
    */
-  private customCatalogApiUrls = new Array<string>();
+  private customProviders = new Array<string>();
 
   constructor(private appConfigService: AppConfigService) {
-    this.presetCatalogApiUrls = this.splitUrls(
+    this.presetProviders = this.splitUrls(
       appConfigService.config.catalogUrl,
     );
   }
@@ -29,24 +30,43 @@ export class CatalogApiUrlService {
    */
   getCatalogApiUrls(): string[] {
     return this.distinct([
-      ...this.presetCatalogApiUrls,
-      ...this.customCatalogApiUrls,
-    ]);
+      ...this.presetProviders,
+      ...this.customProviders,
+    ]).map(url => this.buildCatalogApiUrl(url));
   }
 
   /**
    * Get preset catalog API URLs
    */
-  getPresetApiUrls(): string[] {
-    return this.presetCatalogApiUrls;
+  getPresetProviders(): string[] {
+    return this.presetProviders;
   }
 
-  setCustomApiUrlString(apiUrlString: string) {
-    this.setCustomApiUrls(this.splitUrls(apiUrlString));
+  setCustomProvidersAsString(apiUrlString: string) {
+    this.setCustomProviders(this.splitUrls(apiUrlString));
   }
 
-  setCustomApiUrls(apiUrls: string[]) {
-    this.customCatalogApiUrls = apiUrls;
+  setCustomProviders(apiUrls: string[]) {
+    this.customProviders = apiUrls;
+  }
+
+  /**
+   * We fetch catalogs by proxying them through our own backend.
+   *
+   * @param catalogUrl target connector url (or legacy full url)
+   * @return full url for fetching catalog
+   */
+  private buildCatalogApiUrl(catalogUrl: string) {
+    const baseUrl = this.appConfigService.config.dataManagementApiUrl;
+
+    // Still support manually built URLs
+    const prefix = `${baseUrl}/catalog?providerUrl=`;
+    if (catalogUrl.startsWith(prefix)) {
+      return catalogUrl;
+    }
+
+    // But prefer just entering connector URLs
+    return `${prefix}${encodeURIComponent(catalogUrl)}`;
   }
 
   private splitUrls(commaJoinedUrls?: string | null): string[] {