+++ CHANGELOG.md
@@ -18,6 +18,8 @@ the detailed section referring to by linking pull requests or issues.
 
 - Dashboard with KPIs and charts.
 - Current Connector Organization & URL on Landing Page / Dashboard
+- Catalog URLs no longer need to be prefixed with own connector. Just entering
+  target Connector URL suffices.
 
 #### Changed